refactor: 測試模組重構與coverage 整合
